api/requirements.txt
# API Requirements
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
requests==2.31.0
urllib3>=2.0.0
openai-whisper==20240930
librosa==0.10.2
soundfile==0.12.1
numpy>=1.19.5
deep-translator>=1.11.4
pydantic==2.6.1
python-jose==3.3.0
python-dotenv==1.0.0
psutil>=5.9.0
torch>=2.0.0